+++ fixtures/process.go
@@ -19,10 +19,10 @@ var _ dogma.ProcessRoot = &ProcessRoot{}
 type ProcessMessageHandler struct {
 	NewFunc                  func() dogma.ProcessRoot
 	ConfigureFunc            func(dogma.ProcessConfigurer)
-	RouteEventToInstanceFunc func(context.Context, dogma.Message) (string, bool, error)
-	HandleEventFunc          func(context.Context, dogma.ProcessRoot, dogma.ProcessEventScope, dogma.Message) error
-	HandleTimeoutFunc        func(context.Context, dogma.ProcessRoot, dogma.ProcessTimeoutScope, dogma.Message) error
-	TimeoutHintFunc          func(m dogma.Message) time.Duration
+	RouteEventToInstanceFunc func(context.Context, dogma.Event) (string, bool, error)
+	HandleEventFunc          func(context.Context, dogma.ProcessRoot, dogma.ProcessEventScope, dogma.Event) error
+	HandleTimeoutFunc        func(context.Context, dogma.ProcessRoot, dogma.ProcessTimeoutScope, dogma.Timeout) error
+	TimeoutHintFunc          func(dogma.XMessage) time.Duration
 }
 
 var _ dogma.ProcessMessageHandler = &ProcessMessageHandler{}
@@ -50,33 +50,33 @@ func (h *ProcessMessageHandler) Configure(c dogma.ProcessConfigurer) {
 }
 
 // RouteEventToInstance returns the ID of the process instance that is targeted
-// by m.
+// by e.
 //
 // If h.RouteEventToInstance is non-nil it returns h.RouteEventToInstance(ctx,
-// m), otherwise it panics.
+// e), otherwise it panics.
 func (h *ProcessMessageHandler) RouteEventToInstance(
 	ctx context.Context,
-	m dogma.Message,
+	e dogma.Event,
 ) (string, bool, error) {
 	if h.RouteEventToInstanceFunc == nil {
 		panic(dogma.UnexpectedMessage)
 	}
 
-	return h.RouteEventToInstanceFunc(ctx, m)
+	return h.RouteEventToInstanceFunc(ctx, e)
 }
 
 // HandleEvent handles an event message that has been routed to this handler.
 //
-// If h.HandleEventFunc is non-nil it returns h.HandleEventFunc(ctx, r, s, m),
+// If h.HandleEventFunc is non-nil it returns h.HandleEventFunc(ctx, r, s, e),
 // otherwise it returns nil.
 func (h *ProcessMessageHandler) HandleEvent(
 	ctx context.Context,
 	r dogma.ProcessRoot,
 	s dogma.ProcessEventScope,
-	m dogma.Message,
+	e dogma.Event,
 ) error {
 	if h.HandleEventFunc != nil {
-		return h.HandleEventFunc(ctx, r, s, m)
+		return h.HandleEventFunc(ctx, r, s, e)
 	}
 
 	return nil
@@ -85,16 +85,16 @@ func (h *ProcessMessageHandler) HandleEvent(
 // HandleTimeout handles a timeout message that has been scheduled with
 // ProcessScope.ScheduleTimeout().
 //
-// If h.HandleTimeoutFunc is non-nil it returns h.HandleTimeoutFunc(ctx, r, s, m),
+// If h.HandleTimeoutFunc is non-nil it returns h.HandleTimeoutFunc(ctx, r, s, t),
 // otherwise it returns nil.
 func (h *ProcessMessageHandler) HandleTimeout(
 	ctx context.Context,
 	r dogma.ProcessRoot,
 	s dogma.ProcessTimeoutScope,
-	m dogma.Message,
+	t dogma.Timeout,
 ) error {
 	if h.HandleTimeoutFunc != nil {
-		return h.HandleTimeoutFunc(ctx, r, s, m)
+		return h.HandleTimeoutFunc(ctx, r, s, t)
 	}
 
 	return nil
@@ -105,7 +105,7 @@ func (h *ProcessMessageHandler) HandleTimeout(
 //
 // If h.TimeoutHintFunc is non-nil it returns h.TimeoutHintFunc(m), otherwise it
 // returns 0.
-func (h *ProcessMessageHandler) TimeoutHint(m dogma.Message) time.Duration {
+func (h *ProcessMessageHandler) TimeoutHint(m dogma.XMessage) time.Duration {
 	if h.TimeoutHintFunc != nil {
 		return h.TimeoutHintFunc(m)
 	}