+++ CHANGELOG.md
@@ -6,8 +6,15 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Semantic Versioning].
 
 <!-- references -->
-[Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
-[Semantic Versioning]: https://semver.org/spec/v2.0.0.html
+
+[keep a changelog]: https://keepachangelog.com/en/1.0.0/
+[semantic versioning]: https://semver.org/spec/v2.0.0.html
+
+## Unreleased
+
+### Added
+
+- Added `Command`, `Event` and `Timeout` as aliases for `Message` in preparation for string message typing
 
 ## [0.11.1] - 2021-03-01
 
@@ -158,7 +165,8 @@ The format is based on [Keep a Changelog], and this project adheres to
 - Initial release
 
 <!-- references -->
-[Unreleased]: https://github.com/dogmatiq/dogma
+
+[unreleased]: https://github.com/dogmatiq/dogma
 [0.1.0]: https://github.com/dogmatiq/dogma/releases/tag/v0.1.0
 [0.2.0]: https://github.com/dogmatiq/dogma/releases/tag/v0.2.0
 [0.3.0]: https://github.com/dogmatiq/dogma/releases/tag/v0.3.0